@@ -1,12 +1,12 @@
-import { Response } from '@angular/http'
+import { HttpErrorResponse } from '@angular/common/http'
 import { Observable } from 'rxjs/Observable'
 import 'rxjs/add/observable/throw'
 
 export class ErrorHandler {
-  static handleError(error: Response | any){
+  static handleError(error: HttpErrorResponse | any){
     let errorMessage: string;
 
-    if(error instanceof Response){
+    if(error instanceof HttpErrorResponse){
       errorMessage = `Erro ${error.status} ao acessar a URL ${error.url} - ${error.statusText}`;
     } else {
       errorMessage = error.toString();
