@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Http, Headers, RequestOptions } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 import { ShoppingCartService } from '../restaurant-detail/shopping-cart/shopping-cart.service';
 import { CartItem } from '../restaurant-detail/shopping-cart/cart-item.model';
@@ -11,7 +11,7 @@ import 'rxjs/add/operator/map';
 export class OrderService {
 
   constructor(private cartService: ShoppingCartService,
-              private http: Http){}
+              private http: HttpClient){}
 
   itemsValue(): number {
     return this.cartService.total();
@@ -34,11 +34,7 @@ export class OrderService {
   }
 
   checkOrder(order: Order): Observable<string> {
-    const headers = new Headers();
-    headers.append('Content-Type', 'application/json');
-    return this.http.post(`${MEAT_API}/orders`, JSON.stringify(order), new RequestOptions({headers: headers}))
-          .map(response => response.json())
-          .map(order => order.id);
+    return this.http.post<Order>(`${MEAT_API}/orders`, order).map(order => order.id);
   }
 
   clear() {
